deployment/scripts/requirements.txt
# Requirements for load_catalog.py script
psycopg2-binary>=2.9.0
